+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -5,4 +5,4 @@
 * Added 
   * Created this separate package for terra-dev-site and documentation.
   * Added data objects and examples for Line, Bar and Paired result graphs.
-
+  * Added data objects and examples for Scatter graph.